+++ src/router/routes.js
@@ -50,8 +50,8 @@ const routes = [
       { path: 'cargos', component: () => import('pages/sistema/CargosPage.vue') },
       { path: 'tipos-documento', component: () => import('pages/sistema/TiposDocumentoPage.vue') },
       { path: 'personas', component: () => import('pages/sistema/PersonasPage.vue') },
-      { path: 'persona/nuevo', component: () => import('pages/sistema/PersonaForm2.vue') },
-      { path: 'persona/editar/:id', component: () => import('pages/sistema/PersonaForm2.vue'), props: true },
+      { path: 'persona/nuevo', component: () => import('src/pages/sistema/PersonaForm.vue') },
+      { path: 'persona/editar/:id', component: () => import('src/pages/sistema/PersonaForm.vue'), props: true },
     ],
     meta: { requiresAuth: true },
   },
